Move FilterGroup AppSelectors down the chain
Before, because all the selected state values were in one file, when
one updated it caused re-renders for all of the filter components. By
moving them down to component they are used in, this is prevented.
